Move CashSelection implementations to workflow module.
