+++ finance/src/main/resources/META-INF/services/net.corda.finance.contracts.asset.AbstractCashSelection
