+++ finance/src/test/kotlin/net/corda/finance/contracts/asset/selection/CashSelectionH2ImplTest.kt
@@ -1,4 +1,4 @@
-package net.corda.finance.contracts.asset.cash.selection
+package net.corda.finance.contracts.asset.selection
 
 import net.corda.core.internal.concurrent.transpose
 import net.corda.core.utilities.OpaqueBytes
@@ -16,7 +16,7 @@ import org.junit.After
 import org.junit.Test
 import java.util.Collections.nCopies
 import kotlin.test.assertNotNull
-
+//net.corda.finance.contracts.asset.cash.selection
 class CashSelectionH2ImplTest {
     private val mockNet = MockNetwork(threadPerNode = true, cordappPackages = listOf("net.corda.finance"))
